Rewrite SettingsModal component to React Hooks and styled-components.
