--- src/Modals.css
@@ -1,24 +1,0 @@
-.modal {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    right: auto;
-    bottom: auto;
-	  background-color: white;
-	  padding: 10px;
-	  margin-right: -50%;
-    transform: translate(-50%, -50%);
-}
-
-.overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: darkgray;
-}
-
-.modal input  { width: 350px; }
-.modal h3     { text-align: center; }
-.modal button { text-align: center }
